% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{df_CH}
\alias{df_CH}
\title{Cool data which is used internally in the package}
\format{
An object of class \code{data.frame} with 83 rows and 16 columns.
}
\usage{
df_CH
}
\description{
Cool data which is used internally in the package
}
\keyword{datasets}
